+++ src/ScottPlot5/ScottPlot5/Primitives/BackgroundStyle.cs
@@ -38,7 +38,7 @@ public class BackgroundStyle : IDisposable
     public void Render(SKCanvas canvas, PixelRect target)
     {
         using SKPaint paint = new() { Color = Color.ToSKColor() };
-        Drawing.Fillectangle(canvas, target, paint);
+        Drawing.FillRectangle(canvas, target, paint);
 
         if (Image is not null)
         {